don't attempt version-number auto-discovery
click's version-number autodiscovery causes the program to crash if used
after bundling with pyinstaller, on all platforms.

this change makes the version numbering explicit within the program, and
has setup.py import it instead to maintain DRY

this fixes the problem on linux, so i expect it will be similarly
successful on windows and macos, but as of this commit it is untested.

also changed some imports to relative.